requirements.txt
# Core dependencies for research paper processing pipeline

# AI API clients
anthropic>=0.40.0  # For Claude API
openai>=1.0.0      # For GPT API

# LangChain for orchestration
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5

# PDF text extraction
pypdf>=4.0.0

# Progress bar
tqdm>=4.66.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Retry logic
tenacity>=8.2.0

# HTTP requests for PubMed
requests>=2.31.0

# AWS SDK (optional, for AWS cloud deployment)
boto3>=1.34.0

# Azure SDK (optional, for Azure cloud deployment)
azure-storage-blob>=12.19.0
azure-communication-email>=1.0.0

# Additional utilities
python-dotenv>=1.0.0  # For loading API keys from .env file
pydantic>=2.0.0       # For structured output validation

# Optional: Dashboard and monitoring
streamlit>=1.30.0     # For monitoring dashboard
plotly>=5.18.0        # For visualizations